Add new Waves5 indicator, new sample data, new docs, include LKN and NLWKN station.
The software for the DWR stations is gradually being changed from Waves21 to Waves5 (Datawell). This changes the definition of the timestamp for the *.his and *GPS.txt files from "reception time at the measuring computer" to "end time of the measuring period".

As a new sample data station FINO1 (16th November 2024) was chosen because there are measurements for all three sensors.

There are new versions of the documentations.

Include new LKN and NLWKN DWR stations.

For stations FN1 and BO1, periods with poor RADAC measurements were set to FQF=4.